refactor: split audio tests into modules
