+++ src/audio/tests/mod.rs
@@ -1,0 +1,6 @@
+mod decode;
+mod fade;
+mod playhead_edges;
+mod player;
+mod routing;
+mod support;
